Remove interface and unused code
