+++ src/LanguageServer/Protocol/Handler/AbstractRefreshQueue.cs
@@ -39,7 +39,7 @@ internal abstract class AbstractRefreshQueue :
         LspWorkspaceRegistrationService lspWorkspaceRegistrationService,
         LspWorkspaceManager lspWorkspaceManager,
         IClientLanguageServerManager notificationManager,
-        IFeatureProviderRefresher providerRefresher)
+        FeatureProviderRefresher providerRefresher)
     {
         _asyncListener = asynchronousOperationListenerProvider.GetListener(GetFeatureAttribute());
         _lspWorkspaceRegistrationService = lspWorkspaceRegistrationService;
