+++ src/LanguageServer/Protocol/Handler/InlayHint/InlayHintRefreshQueueFactory.cs
@@ -17,7 +17,7 @@ internal sealed class InlayHintRefreshQueueFactory(
     IAsynchronousOperationListenerProvider asynchronousOperationListenerProvider,
     LspWorkspaceRegistrationService lspWorkspaceRegistrationService,
     IGlobalOptionService globalOptionService,
-    IFeatureProviderRefresher providerRefresher) : ILspServiceFactory
+    FeatureProviderRefresher providerRefresher) : ILspServiceFactory
 {
     public ILspService CreateILspService(LspServices lspServices, WellKnownLspServerKinds serverKind)
     {
